requirements.txt
beautifulsoup4
ffmpy
google
google-cloud-storage
google-cloud-bigquery
PyYAML
soupsieve